Updating decrease_posts_count to prevent negative post count value
If a member only has posts in moderation and none approved. There is
the possibility that if a moderator disapproves their posts, that the
member's post count could become negative.

If this happens, it will cause the check constraint on posts_count >= 0
to fail and an IntegrityError to be raised.

Adding a check that posts_count is greater than 0 before reducing it by one.
If post_count is 0, there is no need to reduce the count.