Add management command to send topic notifications on email
The command can be used in a cron job and should by default be executed
every 5. minute. The interval can be overriden by supplying the argument
`--interval=[in seconds]` as an option. The command iterates through
each newly created post, within the given interval, and sends out
notifications to users that have enabled receiving notifications on
subscribed topics.