Take admin permissions into account for user/group permission forms on forums
References #126.

The forms to select user or group to edit it's permissions (on a forum)
are now only built if you have the right (granted via the auth system) to change
the user or group permissions on forums.

Possible errors (not choosing ID/anonymous user but still submitting) are
displayed above the form that was actually submitted.

In the template the html is only made for the forms that need to be shown. And
for the group permission form a section for errors is added.

NOTE: the user still needs permission to change (i.e. see) users and groups to
actually be able to use these forms (the selection of the id), this has not
changed.