Add setting to define the widget to use for selecting user/group for permissions
References #126

REASON:
Before this commit the widget that is used to select a user or group to alter
its permissions is Django's ForeignKeyRawIdWidget. This widget requires that
the user has change right on the group and user to actually see the list to
select the target from.
Also, the widget presents the user with an integer field which might be
perceived as not completely user friendly.
This approach was taken to reduce the load associated with presenting a select
box of targets.

PROPOSITION:
Now the developer can decide what widget to use to select user or group by
writing a callable and putting the dotted path to it inside the new settings.
This may take away the need to have to grant change right on users/groups.
This also provides the possibility to write a callable that influences the
presentation (for example full name instead of user name for user Select).
When the setting is not overwritten in own settings file the default is still
used.

CHANGES:
Added setting for both group and user (though default these both use the same
callable).
Added documentation for these settings.
Created callable to link to in setting, this callable just returns the
ForeignKeyRawIdWidget like in the former behavior.
Add helperfunction _get_setting_callable to get an actual callable from the
dotted path in the setting.
Use the new settings in PickUserForm and PickGroupForm.