Fix test on redirection to group/user/anonymous_user permission forms
editpermissions_index_view in admin.py checks which form is posted by checking
the presence of '_select_user' or '_select_group' (save button) in the post and
will not return a HttpResponseObject if the form was not posted by 'use' of a
button.
In the test we now simulate the use of the corresponding button by passing on
the parameter manually.