Use more unique name for factories forums
In test output it is unhelpful for debugging if all forums have the same name.
So we now pick a random name from a list of (hopefully but not guaranteed)
unique names. Instead of having all forums with the same name we now have a big
chance (but not 100%) that all forums have a unique name. And also a short
(person) name at that, instead of a long string that clogs up the output.