Refactor code to prevent duplicate SQL queries
Rewrite the code so that we do not have to repeat queries for every item in a
loop.
There's a new helper function get_perms_for_forumlist in checker.py that
returns a dictionary of forum to permissions for the user, taking a list of
forums and optionally a list of permission codenames as arguments.
Both get_perms in chekcer.py and _get_forums_for_user in handler.py use this
function.
This way the computing of which permissions are in effect is still done in one
place making it easier to maintain and add to and consistent for both
handler.py and checker.py.