Fix comparison for more strict databases
Travis gave failures on tests because a comparison of two ids was done with
'is' instead of '=='. sqlite, which is by default used for testing apparently
returns ids from the database in a comparable form so this 'is' comparison
worked, but mysql (and assumed also postgresql) are stricter in the type they
return. So comparison is now done by '==' resulting in all tests to pass again.