Target all loggedin users (#170)
* Have handler._get_forums_for_user use code from checker.py

Since both handler._get_forums_for_user and checker.get_perms have largly the
same code to check permissions, have the handler function now use the checker
function so we have this code just in one place.

* Add authenticated_user as a target option to AbstractUserForumPermission

Make it possible to target all authenticated users (as opposed to a specific
one on ID or the anonymous user). So we add this field to the model and update
the validation.
This also requires a migration file.

* Add admin interface for targeting authenticated users

Add interface to admin to choose the authenticated user as target for
permissions. Also add template and form.

* In UserForumPermissionAdmin show if authenticated user was targeted

Add a column with boolean value to show if the authenticated user (as opposed
to one specific user or anonymous user) is targeted for a permission.

* Have get_perms take into account the authenticated_users target for permissions

Add code to collect the permissions that were set for all authenticated users
but take into account that it has the least precedence (group and specific user
are more specific targets).

* Add tests for new permission-target all authenticated users

Add two tests to check that permissions on all authenticated users have less
precedence (and do work) than group/specific-user permissions.
In shortcuts.py we edit assign_perm and remove_perm to build a specific query
when the target authenticated_users is used.

* Add documentation for new permission target

Document the extra permission target; all authenticated users.

* Add tests for new admin interfaces around authenticated users

With the new permission target (all authenticated users) we need to test if the
admin also works for this new target of the PickUserForm.

* Change formatting of long texts and fix one faulty comment

Change formatting (according to request) for long lines and fix a copy-paste
error in a comment so that it is now correct.

* Use more unique name for factories forums

In test output it is unhelpful for debugging if all forums have the same name.
So we now pick a random name from a list of (hopefully but not guaranteed)
unique names. Instead of having all forums with the same name we now have a big
chance (but not 100%) that all forums have a unique name. And also a short
(person) name at that, instead of a long string that clogs up the output.

* Refactor code to prevent duplicate SQL queries

Rewrite the code so that we do not have to repeat queries for every item in a
loop.
There's a new helper function get_perms_for_forumlist in checker.py that
returns a dictionary of forum to permissions for the user, taking a list of
forums and optionally a list of permission codenames as arguments.
Both get_perms in chekcer.py and _get_forums_for_user in handler.py use this
function.
This way the computing of which permissions are in effect is still done in one
place making it easier to maintain and add to and consistent for both
handler.py and checker.py.

* Fix comparison for more strict databases

Travis gave failures on tests because a comparison of two ids was done with
'is' instead of '=='. sqlite, which is by default used for testing apparently
returns ids from the database in a comparable form so this 'is' comparison
worked, but mysql (and assumed also postgresql) are stricter in the type they
return. So comparison is now done by '==' resulting in all tests to pass again.

* Change name of constant to something more descriptive

* Optimize query filters

Combine two filters in one query and do the actual filtering later in python so
the amount of necessary queries is lessened.