Observe admin permissions on forum permissions (#172)
* Take admin permissions into account for user/group permission forms on forums

References #126.

The forms to select user or group to edit it's permissions (on a forum)
are now only built if you have the right (granted via the auth system) to change
the user or group permissions on forums.

Possible errors (not choosing ID/anonymous user but still submitting) are
displayed above the form that was actually submitted.

In the template the html is only made for the forms that need to be shown. And
for the group permission form a section for errors is added.

Also add tests to check if the forms will give an error when a choice is
omitted but 'submit' is pressed.

NOTE: the user still also needs permission to view users and groups to
actually be able to use these forms (the selection of the id), this has not
changed.

* Fix test on redirection to group/user permission forms

editpermissions_index_view in admin.py checks which form is posted by checking
the presence of '_select_user' or '_select_group' (save button) in the post and
will not return a HttpResponseObject if the form was not posted by 'use' of a
button.
In the test we now simulate the use of the corresponding button by passing on
the parameter manually.