Show users and groups that already have permissions on a forum
In the index page of permissions for a specific forum (choosing which group/user
to go edit the permissions for) show (in text) which users and groups already
have permissions set for this forum.
This makes is easier to select a user/group to edit existing permissions for
and it also reminds the admin user which users/groups he has already made
permissions for for the chosen forum.