Also show anonymous/authenticated for already set user permissions
Next to specific users there can also be permissions set for anonymous or all
authenticated users. So we want to display this (instead of None None (None)).
For this it is easier to make a list with all display strings in it to pass on
to the template.
We now also ask for the queryset instead of certain values and can no longer
use 'distinct'. We take care of that now in python code instead of in the query.