Add code to test of view to keep coverage
The newly added code for displaying which users/groups already have permissions
on a forum was not completely covered because the current tests did not result
in this code being run. To make sure it is run we set some permissions in
test_editpermission_index_view_browsing_works that actually tests if the view
does not break.