Move overview of users/groups with permissions to own view
Instead of showing the users/groups with permissions inside the select
user/group forms of the editpermissions_index, we now put that in it's own view
and template and add a link to each form to open this as a popup.

This keeps the index view concise but still makes the feature readily available.