Also allow for global permissions
Oops... it should also be possible to check which groups/users already have
global permissions. So we need to have 2 possible urls.
To prevent two if/else blocks in the template we pass on the url for the link
(with/without forum id) from the view to the template.

Also adjust test to check if both links work and if the page they should call
up works.