Fix bug of all_authenticated_users permission not working without group
When a logged in user does not have a group with rights on forum X, the
all_authenticated_users permissions for forum X were not taken into account.

Add a test to prove this bug and fix it by taking a block of code outside of
the if that checks if there were group permissions on the given forum, but keep
it within the if for not self.user.is_anonymous.

Closes #193