Move final computation of permcodes to bottom of function
To be able to take into account the per_forum_nongranted_all_users_permcodes
when making the final granted_user_permcodes list, we need to move the block
that computers this list to the bottom of the whole get_perms_for_forumlist
function.

We also add a test to ensure that all_authenticated_users permissions take
precedence over the default_authenticated_permissions (which possibly come from
settings), which is now the case by doing to code move mentioned above.