Make a test more clear by using a different permission to check
The test class sets the can_see_forum permission in
DEFAULT_AUTHENTICATED_USER_FORUM_PERMISSIONS at setup. To make the test
test_knows_that_alluser_permissions_take_precedence_over_alluser_global_permissions
more transparent we do not use can_see_forum to check here but
can_edit_own_posts. Effectively ignoring (not using) the default permissions
setting for this test.

Using can_see_forum obscured the bug in #193 because the default permissions
set it to True, passing the check in this test which would not have passed with
a different permission to check that was not elsewhere set to True.