Fix buggy allauthusers permissions (#194)
* Fix bug of all_authenticated_users permission not working without group

When a logged in user does not have a group with rights on forum X, the
all_authenticated_users permissions for forum X were not taken into account.

Add a test to prove this bug and fix it by taking a block of code outside of
the if that checks if there were group permissions on the given forum, but keep
it within the if for not self.user.is_anonymous.

Closes #193

* Move final computation of permcodes to bottom of function

To be able to take into account the per_forum_nongranted_all_users_permcodes
when making the final granted_user_permcodes list, we need to move the block
that computers this list to the bottom of the whole get_perms_for_forumlist
function.

We also add a test to ensure that all_authenticated_users permissions take
precedence over the default_authenticated_permissions (which possibly come from
settings), which is now the case by doing to code move mentioned above.

* Make a test more clear by using a different permission to check

The test class sets the can_see_forum permission in
DEFAULT_AUTHENTICATED_USER_FORUM_PERMISSIONS at setup. To make the test
test_knows_that_alluser_permissions_take_precedence_over_alluser_global_permissions
more transparent we do not use can_see_forum to check here but
can_edit_own_posts. Effectively ignoring (not using) the default permissions
setting for this test.

Using can_see_forum obscured the bug in #193 because the default permissions
set it to True, passing the check in this test which would not have passed with
a different permission to check that was not elsewhere set to True.